refactor(wip): moved leaderboard into redux to prevent rerender of entire component tree
