+++ src/redux/informationSlice.ts
@@ -1,6 +1,19 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 
+export interface LeaderboardBody {
+    index: number;
+    mass: number;
+    color: string;
+    dOrigin: number;
+    dTarget: number;
+    orbiting: number;
+    dOrbit: number;
+    orbitColor: string;
+    numSatellites: number;
+}
+
 export interface InformationState {
+    leaderboard: Array<LeaderboardBody>;
     numActiveBodies: number;
     numStars: number;
     yearsElapsed: number;
@@ -8,6 +21,7 @@ export interface InformationState {
 }
 
 const initialState: InformationState = {
+    leaderboard: [],
     numActiveBodies: 0,
     numStars: 0,
     yearsElapsed: 0,
@@ -18,6 +32,9 @@ export const informationSlice = createSlice({
     name: "information",
     initialState,
     reducers: {
+        setLeaderboard: (state, action: PayloadAction<Array<LeaderboardBody>>) => {
+            state.leaderboard = action.payload;
+        },
         setNumActiveBodies: (state, action: PayloadAction<number>) => {
             state.numActiveBodies = action.payload;
         },
